requirements.txt
python-mpd2
tabulate
